Create a through-table for indexed resources containing text extraction data
