+++ solution/backend/content_search/text_extraction.py
