+++ solution/backend/resources/urls.py
@@ -2,7 +2,6 @@ from django.urls import path
 
 from resources.views import (
     CitationViewSet,
-    ContentTextViewSet,
     FederalRegisterLinksNumberViewSet,
     FederalRegisterLinkViewSet,
     InternalCategoryViewSet,
@@ -26,7 +25,6 @@ urlpatterns = [
         "get": "list",
     })),
     path("<int:id>/edit", ResourceEditViewSet.as_view(), name="edit"),
-    path("<int:id>/content", ContentTextViewSet.as_view(), name="content"),
     path("public", PublicResourceViewSet.as_view({
         "get": "list",
     })),
