+++ solution/backend/resources/views/__init__.py
@@ -4,4 +4,3 @@ from .groups import *  # noqa
 from .subjects import *  # noqa
 from .resources import *  # noqa
 from .redirects import *  # noqa
-from .text_extraction import *  # noqa
